requirements.txt
Django==5.2.8
djangorestframework==3.15.2
drf-spectacular==0.27.2
pytest==8.3.3
pytest-django==4.9.0
python-dotenv==1.0.1
flake8==7.1.1
psycopg2-binary==2.9.9
gunicorn==23.0.0

